fix(finance): penalites de retard et montants dus en Decimal (suite #661)
Les relances de paiement portaient des sommes reclamees a un coproprietaire
— dont des penalites au taux legal civil belge — en DOUBLE PRECISION / f64,
de la colonne SQL jusqu'au DTO. Defaut revele par le gate anti-f64 livre
avec #661, ou il avait ete gele en dette tracee.

- Migration 20260826000000 : amount_owed, penalty_amount, total_amount en
  NUMERIC(12,2).
- La contrainte `CHECK (total_amount = amount_owed + penalty_amount)` etait
  une egalite flottante EXACTE : elle pouvait rejeter une ligne valide de
  facon non reproductible. Elle est desormais fiable.
- `calculate_penalty` : l'arrondi `(x * 100.0).round() / 100.0` en f64 est
  remplace par un arrondi COMMERCIAL au centime (MidpointAwayFromZero), et
  non le banker's rounding que `round_dp` applique par defaut.
- L'invariant « au moins un centime », porte par `#[validate(range)]` sur le
  DTO (inapplicable a un Decimal), descend dans `PaymentReminder::new` : il
  vaut desormais pour tous les appelants, pas seulement la route HTTP.
- Suppression d'un aller-retour Decimal->f64 sur `expenses.amount`, colonne
  pourtant deja NUMERIC.
- Assertions de test en egalite exacte : le step BDD tolerait jusqu'ici
  UN EURO d'ecart sur une penalite de retard legale.
- Les 5 entrees payment_reminder sortent de l'allowlist du gate.

Migration rejouee sur une base reconstruite depuis zero, ce qui a revele
deux bugs qu'aucune suite de tests n'aurait signales :
  1. le DROP CONSTRAINT visait des noms inventes (PostgreSQL avait nomme
     les CHECK anonymes `payment_reminders_check`/`check2`) : la contrainte
     survivait et une redondante etait creee ;
  2. `ALTER COLUMN TYPE` ne reecrit pas un CHECK, il y insere un cast vers
     l'ancien type — la positivite serait restee evaluee en flottant.
Les 6 contraintes de la table sont desormais evaluees en NUMERIC.

Verifie : cargo check --tests propre, 1669 tests unitaires verts (13/13 sur
payment_reminder), clippy -D warnings exit 0, fmt propre, migration validee
sur base neuve.

Refs: #661

+++ backend/src/application/dto/payment_reminder_dto.rs
@@ -1,4 +1,5 @@
 use crate::domain::entities::{DeliveryMethod, PaymentReminder, ReminderLevel, ReminderStatus};
+use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 use validator::Validate;
 
@@ -10,8 +11,15 @@ pub struct CreatePaymentReminderDto {
     pub owner_id: String,
     pub level: ReminderLevel,
 
-    #[validate(range(min = 0.01))]
-    pub amount_owed: f64,
+    /// Montant dû en euros. `Decimal` exact (ADR-0007/0008, suite #661).
+    ///
+    /// Plus d'annotation `#[validate(range)]` : `validator` ne sait pas borner
+    /// un `Decimal`. L'invariant n'est pas perdu pour autant — il est porté par
+    /// `PaymentReminder::new`, qui rejette tout montant inférieur au centime.
+    /// C'est aussi la place correcte au regard de l'architecture hexagonale, et
+    /// le pattern déjà suivi par `expense_dto` / `budget_dto` pour leurs
+    /// montants.
+    pub amount_owed: Decimal,
 
     pub due_date: String, // ISO 8601 format
 
@@ -30,9 +38,9 @@ pub struct PaymentReminderResponseDto {
     pub owner_email: Option<String>, // Owner email for contact
     pub level: ReminderLevel,
     pub status: ReminderStatus,
-    pub amount_owed: f64,
-    pub penalty_amount: f64,
-    pub total_amount: f64,
+    pub amount_owed: Decimal,
+    pub penalty_amount: Decimal,
+    pub total_amount: Decimal,
     pub due_date: String,
     pub days_overdue: i64,
     pub delivery_method: DeliveryMethod,
@@ -102,8 +110,8 @@ pub struct AddTrackingNumberDto {
 /// DTO for payment recovery dashboard statistics
 #[derive(Debug, Serialize, Clone)]
 pub struct PaymentRecoveryStatsDto {
-    pub total_owed: f64,
-    pub total_penalties: f64,
+    pub total_owed: Decimal,
+    pub total_penalties: Decimal,
     pub reminder_counts: Vec<ReminderLevelCountDto>,
     pub status_counts: Vec<ReminderStatusCountDto>,
 }
@@ -126,13 +134,13 @@ pub struct OverdueExpenseDto {
     pub expense_id: String,
     pub owner_id: String,
     pub days_overdue: i64,
-    pub amount: f64,
+    pub amount: Decimal,
     pub recommended_level: ReminderLevel,
 }
 
 impl OverdueExpenseDto {
     /// Create DTO with automatically determined reminder level
-    pub fn new(expense_id: String, owner_id: String, days_overdue: i64, amount: f64) -> Self {
+    pub fn new(expense_id: String, owner_id: String, days_overdue: i64, amount: Decimal) -> Self {
         let recommended_level = if days_overdue >= 60 {
             ReminderLevel::FormalNotice
         } else if days_overdue >= 30 {
